requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
dataclasses-json==0.5.6
distlib @ file:///tmp/build/80754af9/distlib_1622544193819/work
eventbrite==3.3.5
facebook-sdk==3.1.0
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
h11==0.12.0
idna==3.3
marshmallow==3.13.0
marshmallow-enum==1.5.1
meetup-api==0.1.1
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector-python==8.0.26
mysqlclient==2.0.3
numpy==1.21.2
nutch==1.10.3
oauthlib==3.1.1
outcome==1.1.0
pandas==1.3.3
protobuf==3.18.1
pycparser==2.20
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-facebook-api==0.11.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
selenium==4.0.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
virtualenv @ file:///C:/ci/virtualenv_1620977881384/work
wincertstore==0.2
wsproto==1.0.0
